20190717|python 함수 정리

+++ 알고리즘공부/SWEA/2063.py
@@ -1,4 +1,4 @@
 num = int(input())
 case = list(map(int, input().split(' ')))
 
-print(sorted(case)[len(case)/2])
+print(sorted(case)[int(len(case)/2)])
